Optimize lock implementation
Avoid the need to allocate lwmutex objects by turning _LOCK_T and
_LOCK_RECURSIVE_T themselves into sys_lwmutex structs.

Since the lock objects are no longer pointers, some knowledge about
the internal structure of a lwmutex is needed to be able to differentiate
one which has been initialized from one which hasn't.  We happen to know
that the attribute field is always nonzero (constant 0x12, actually)
when the lwmutex is initialized, so we'll use that.